requirements.txt
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
mysql-connector-python==8.3.0